Add export FHIR resources to GCS IO connector.
