+++ sdks/java/io/google-cloud-platform/src/test/java/org/apache/beam/sdk/io/gcp/healthcare/FhirIOTest.java
@@ -17,16 +17,10 @@
  */
 package org.apache.beam.sdk.io.gcp.healthcare;
 
-import static org.apache.beam.sdk.io.gcp.healthcare.FhirIOTestUtil.DEFAULT_TEMP_BUCKET;
-import static org.apache.beam.sdk.io.gcp.healthcare.HL7v2IOTestUtil.HEALTHCARE_DATASET_TEMPLATE;
-
-import java.io.IOException;
-import java.security.SecureRandom;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import org.apache.beam.sdk.io.TextIO;
 import org.apache.beam.sdk.testing.PAssert;
 import org.apache.beam.sdk.testing.TestPipeline;
 import org.apache.beam.sdk.transforms.Count;
@@ -35,10 +29,7 @@ import org.apache.beam.sdk.transforms.MapElements;
 import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.PCollection;
 import org.apache.beam.sdk.values.TypeDescriptors;
-import org.junit.After;
-import org.junit.AfterClass;
 import org.junit.Assert;
-import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -47,60 +38,6 @@ import org.junit.runners.JUnit4;
 @RunWith(JUnit4.class)
 public class FhirIOTest {
   @Rule public final transient TestPipeline pipeline = TestPipeline.create();
-
-  private final String fhirStoreId;
-  private final String project;
-  private transient HealthcareApiClient client;
-  private String healthcareDataset;
-  private String version;
-
-  public FhirIOTest() {
-    long testTime = System.currentTimeMillis();
-    this.fhirStoreId = "FHIR_store_" + testTime + "_" + (new SecureRandom().nextInt(32));
-    this.version = "STU3";
-    this.project =
-        TestPipeline.testingPipelineOptions()
-            .as(HealthcareStoreTestPipelineOptions.class)
-            .getStoreProjectId();
-  }
-
-  @Before
-  public void setup() throws Exception {
-    healthcareDataset = String.format(HEALTHCARE_DATASET_TEMPLATE, project);
-    if (client == null) {
-      this.client = new HttpHealthcareApiClient();
-    }
-    client.createFhirStore(healthcareDataset, fhirStoreId, version, null);
-
-    // Execute bundles to populate some data.
-    FhirIOTestUtil.executeFhirBundles(
-        client,
-        healthcareDataset + "/fhirStores/" + fhirStoreId,
-        FhirIOTestUtil.BUNDLES.get(version));
-  }
-
-  @After
-  public void deleteFHIRtore() throws IOException {
-    HealthcareApiClient client = new HttpHealthcareApiClient();
-    client.deleteFhirStore(healthcareDataset + "/fhirStores/" + fhirStoreId);
-  }
-
-  @AfterClass
-  public static void teardownBucket() throws IOException {
-    FhirIOTestUtil.tearDownTempBucket();
-  }
-
-  @Test
-  public void test_FhirIO_exportFhirResourcesGcs() {
-    String fhirStoreName = healthcareDataset + "/fhirStores/" + fhirStoreId;
-    String exportGcsUriPrefix =
-        "gs://" + DEFAULT_TEMP_BUCKET + "/" + (new SecureRandom().nextInt(32));
-    FhirIO.ExportGcs.Result exportResults =
-        pipeline.apply(FhirIO.exportResourcesToGcs(fhirStoreName, exportGcsUriPrefix));
-    PCollection<String> resources = exportResults.getResources();
-    resources.apply(TextIO.write().to(exportGcsUriPrefix + "/write"));
-    pipeline.run();
-  }
 
   @Test
   public void test_FhirIO_failedReads() {
@@ -153,16 +90,16 @@ public class FhirIOTest {
 
   private static ArrayList<KV<String, String>> createTestData() {
     String[] scientists = {
-      "Einstein",
-      "Darwin",
-      "Copernicus",
-      "Pasteur",
-      "Curie",
-      "Faraday",
-      "Newton",
-      "Bohr",
-      "Galilei",
-      "Maxwell"
+        "Einstein",
+        "Darwin",
+        "Copernicus",
+        "Pasteur",
+        "Curie",
+        "Faraday",
+        "Newton",
+        "Bohr",
+        "Galilei",
+        "Maxwell"
     };
     ArrayList<KV<String, String>> data = new ArrayList<>();
     for (int i = 0; i < NUM_ELEMENTS; i++) {
